squashed more helper funcs
